Code improve
