+++ utils/locale.ts
@@ -1,3 +1,6 @@
+import { match } from '@formatjs/intl-localematcher';
+import Negotiator from 'negotiator';
+import { NextRequest } from 'next/server';
 import { DEFAULT_LOCALE, SUPPORTED_LOCALES } from '../middleware';
 
 export function saveLocaleToCookie(locale: string) {
@@ -11,4 +14,40 @@ export function saveLocaleToCookie(locale: string) {
 
   document.cookie = `NEXT_LOCALE=${locale}; expires=${expiryDate.toUTCString()}; path=/; SameSite=Lax`;
   console.log(`Saved locale to cookie: ${locale}`);
+}
+
+export function getLocale(request: NextRequest): string {
+  const cookieLocale = getLocaleFromCookie(request);
+  if (cookieLocale) {
+    return cookieLocale;
+  }
+
+  const acceptLanguageLocale = getLocaleFromAcceptLanguage(request);
+  if (acceptLanguageLocale) {
+    return acceptLanguageLocale;
+  }
+
+  return DEFAULT_LOCALE;
+}
+
+function getLocaleFromCookie(request: NextRequest): string | null {
+  const cookieLocale = request.cookies.get('NEXT_LOCALE')?.value;
+
+  if (cookieLocale && SUPPORTED_LOCALES.includes(cookieLocale)) {
+    return cookieLocale;
+  }
+
+  return null;
+}
+
+function getLocaleFromAcceptLanguage(request: NextRequest): string | null {
+  const negotiatorHeaders: Record<string, string> = {};
+  request.headers.forEach((value, key) => (negotiatorHeaders[key] = value));
+
+  let languages = new Negotiator({ headers: negotiatorHeaders }).languages();
+  try {
+    return match(languages, SUPPORTED_LOCALES, DEFAULT_LOCALE);
+  } catch (error) {
+    return null;
+  }
 }
